+++ Sources/SuperwallKit/Analytics/Internal Tracking/Trackable Events/TrackableSuperwallEvent.swift
@@ -416,6 +416,7 @@ enum InternalSuperwallEvent {
       if let demandTier = demandTier {
         params["attr_demandTier"] = demandTier
       }
+      params["user_attributes"] = Superwall.shared.userAttributes
       return params
     }
     var audienceFilterParams: [String: Any] {
@@ -630,10 +631,7 @@ enum InternalSuperwallEvent {
         if let demandTier = demandTier {
           placementParams["attr_demandTier"] = demandTier
         }
-        let appleSearchAttributes = Superwall.shared.userAttributes.filter {
-          $0.key.hasPrefix("apple_search_ads_")
-        }
-        placementParams += appleSearchAttributes
+        placementParams["user_attributes"] = Superwall.shared.userAttributes
         fallthrough
       case .start,
         .abandon,
@@ -776,6 +774,7 @@ enum InternalSuperwallEvent {
       case fail(Error)
       case complete
       case retry(Int)
+      case missingProducts(Set<String>)
     }
     let state: State
     var audienceFilterParams: [String: Any] {
@@ -798,6 +797,12 @@ enum InternalSuperwallEvent {
           paywallInfo: paywallInfo,
           attempt: attempt
         )
+      case .missingProducts(let identifiers):
+        return .paywallProductsLoadMissingProducts(
+          triggeredPlacementName: placementData?.name,
+          paywallInfo: paywallInfo,
+          identifiers: identifiers
+        )
       }
     }
     let paywallInfo: PaywallInfo
@@ -810,6 +815,9 @@ enum InternalSuperwallEvent {
       ]
       if case .fail(let error) = state {
         params["error_message"] = error.safeLocalizedDescription
+      }
+      if case .missingProducts(let identifiers) = state {
+        params["missing_products"] = Array(identifiers).joined(separator: ",")
       }
       params += await paywallInfo.placementParams()
       return params
@@ -998,6 +1006,22 @@ enum InternalSuperwallEvent {
       return [
         "request_url": requestURLString,
         "response": responseString
+      ]
+    }
+  }
+
+  struct ReviewRequested: TrackableSuperwallEvent {
+    let count: Int
+    let type: ReviewType
+    var superwallEvent: SuperwallEvent {
+      return .reviewRequested(count: count)
+    }
+    var audienceFilterParams: [String: Any] = [:]
+
+    func getSuperwallParameters() async -> [String: Any] {
+      return [
+        "count": count,
+        "type": type.rawValue
       ]
     }
   }