+++ Sources/SuperwallKit/Analytics/Superwall Placement/SuperwallEventObjc.swift
@@ -131,6 +131,9 @@ public enum SuperwallEventObjc: Int, CaseIterable {
   /// When the request to load the paywall's products has failed and is being retried.
   case paywallProductsLoadRetry
 
+  /// When the paywall's products are missing from the App Store.
+  case paywallProductsLoadMissingProducts
+
   /// Information about a paywall presentation request
   case paywallPresentationRequest
 
@@ -211,6 +214,9 @@ public enum SuperwallEventObjc: Int, CaseIterable {
   /// When the integration attributes are set.
   case integrationAttributes
 
+  /// When a review is requested from the user.
+  case reviewRequested
+
   public init(event: SuperwallEvent) {
     self = event.backingData.objcEvent
   }
@@ -289,6 +295,8 @@ public enum SuperwallEventObjc: Int, CaseIterable {
       return "paywallProductsLoad_fail"
     case .paywallProductsLoadRetry:
       return "paywallProductsLoad_retry"
+    case .paywallProductsLoadMissingProducts:
+      return "paywallProductsLoad_missingProducts"
     case .paywallProductsLoadComplete:
       return "paywallProductsLoad_complete"
     case .paywallPresentationRequest:
@@ -343,6 +351,8 @@ public enum SuperwallEventObjc: Int, CaseIterable {
       return "networkDecoding_fail"
     case .integrationAttributes:
       return "integration_attributes"
+    case .reviewRequested:
+      return "review_requested"
     }
   }
 }