+++ Sources/SuperwallKit/Paywall/View Controller/PaywallViewController.swift
@@ -109,7 +109,7 @@ public class PaywallViewController: UIViewController, LoadingDelegate {
   private var popupHeightConstraint: NSLayoutConstraint?
   var popupContainerView: UIView?
 
-  /// Internal property for transition logic
+  /// Internal property for transition logic testing
   var isCustomBackgroundDismissal = false
 
   /// The background color of the paywall, depending on whether the device is in dark mode.
@@ -986,6 +986,7 @@ extension PaywallViewController: PaywallMessageHandlerDelegate {
       } else {
         SKStoreReviewController.requestReview()
       }
+      trackReviewRequest(type: .inApp)
     case .external:
       let appId: String
       if let iosAppId = factory.makeAppId() {
@@ -1003,7 +1004,19 @@ extension PaywallViewController: PaywallMessageHandlerDelegate {
 
       if let url = URL(string: "https://apps.apple.com/app/id\(appId)?action=write-review") {
         UIApplication.shared.open(url)
+        trackReviewRequest(type: .external)
       }
+    }
+  }
+
+  private func trackReviewRequest(type: ReviewType) {
+    Task {
+      let count = await deviceHelper.reviewRequestsTotal() + 1
+      let reviewRequestEvent = InternalSuperwallEvent.ReviewRequested(
+        count: count,
+        type: type
+      )
+      await Superwall.shared.track(reviewRequestEvent)
     }
   }
 }